+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `\Heptacom\HeptaConnect\Storage\Base\Contract\ReceptionRouteListResult\RouteFindByTargetsAndTypeActionInterface` for checking the existence of a route by its components by the given `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteFindByTargetsAndTypeCriteria` to return a `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteFindByTargetsAndTypeResult`
 - Add `\Heptacom\HeptaConnect\Storage\Base\Contract\ReceptionRouteListResult\RouteGetActionInterface` for reading metadata of routes by the given `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteGetCriteria` to return a `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteGetResult`
 - Add `\Heptacom\HeptaConnect\Storage\Base\Contract\ReceptionRouteListResult\RouteCreateActionInterface` for creating routes by the given `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteCreateParams` and `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteCreateParam` to return a `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteCreateResult`
+- Add `\Heptacom\HeptaConnect\Storage\Base\Contract\ReceptionRouteListResult\RouteCapabilityOverviewActionInterface` for listing available route capabilities by the given `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteCapabilityOverviewCriteria` to return a `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteCapabilityOverviewResult`
 
 ### Changed